requirements.txt
Flask-Cors==2.1.2
flask-restplus==0.9.2
tinydb==3.2.1
